# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing caviar products and orders",
        "title": "Caviar API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product with variants and details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "parameters": [
                    {
                        "description": "Product creation data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.ProductCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created product",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.ProductCreateDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/products/search": {
            "post": {
                "description": "Search and filter products with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Search products",
                "parameters": [
                    {
                        "description": "Product filter parameters",
                        "name": "filter",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_types.ProductFilter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/caviar_internal_dto.ProductResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update product details, variants, and other information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product update data",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.ProductUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated product",
                        "schema": {
                            "$ref": "#/definitions/caviar_internal_dto.ProductUpdateDTO"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a product and all its variants",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "caviar_internal_dto.CaviarDetailsDTO": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "fishAge": {
                    "type": "string"
                },
                "grainSize": {
                    "type": "string"
                },
                "shelfLife": {
                    "$ref": "#/definitions/caviar_internal_dto.ShelfLifeDTO"
                },
                "taste": {
                    "type": "string"
                },
                "texture": {
                    "type": "string"
                }
            }
        },
...
        "caviar_internal_dto.MoneyDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                }
            }
        },
...
        "caviar_internal_dto.ProductCreateDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "details": {
                    "$ref": "#/definitions/caviar_internal_dto.CaviarDetailsDTO"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.VariantCreateDTO"
                    }
                }
            }
        },
        "caviar_internal_dto.ProductResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "details": {
                    "$ref": "#/definitions/caviar_internal_dto.CaviarDetailsDTO"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.VariantResponseDTO"
                    }
                }
            }
        },
        "caviar_internal_dto.ProductUpdateDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "details": {
                    "$ref": "#/definitions/caviar_internal_dto.CaviarDetailsDTO"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/caviar_internal_dto.VariantUpdateDTO"
                    }
                }
            }
        },
...
        "caviar_internal_dto.ShelfLifeDTO": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "string"
                },
                "tempRange": {
                    "$ref": "#/definitions/caviar_internal_dto.TemperatureRangeDTO"
                }
            }
        },
        "caviar_internal_dto.TemperatureRangeDTO": {
            "type": "object",
            "properties": {
                "maxC": {
                    "type": "number"
                },
                "minC": {
                    "type": "number"
                }
            }
        },
        "caviar_internal_dto.VariantCreateDTO": {
            "type": "object",
            "properties": {
                "mass": {
                    "type": "integer"
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                    }
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "caviar_internal_dto.VariantResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mass": {
                    "type": "integer"
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                    }
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "caviar_internal_dto.VariantUpdateDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mass": {
                    "type": "integer"
                },
                "prices": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/caviar_internal_dto.MoneyDTO"
                    }
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "caviar_internal_types.ProductFilter": {
            "type": "object",
            "properties": {
                "createdAfter": {
                    "type": "string"
                },
                "createdBefore": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "includeVariants": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "offset": {
                    "type": "integer"
                },
                "search": {
                    "type": "string"
                },
                "showAll": {
                    "type": "boolean"
                },
                "slug": {
                    "type": "string"
                },
                "sortBy": {
                    "type": "string"
                },
                "sortOrder": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "string"
                },
                "updatedAfter": {
                    "type": "string"
                },
                "updatedBefore": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}